Updated openstack/openstack Project: openstack/nova  30599a5c4f534f0d225273035c39489b113d3558
Hyper-V: Fix SMBFS volume attach race condition

When the SMBFS volume backend is used and a volume is detached,
the according SMB share is unmounted if no longer used.

This can cause issues if at the same time, a different volume
stored on the same share is being attached as the according
disk image will not be available.

This patch fixes the issue by synchronizing volume attach and share
unmount operations based on the share path.

Closes-Bug: #1461081

Change-Id: I45f202ad6ada0c472759e8ba3c122f3cfbeee7cc