Updated openstack/openstack Project: openstack/nova  0142fc20f14563ff871733955424a5e9a61b2e4d
Deprecate the osapi_v3.enabled option

This is part of the larger effort in Liberty to remove references to the
v3 API.

This enables the v2.1 API by default and deprecates the 'enabled' option
for removal in the N release.

Alternatively we could just ignore the option altogether but changing it
to True by default essentially does that for us.

Related-Bug: #1462901

Change-Id: I43f0352f9fa89401f79389a6dc1035d901f52ed2