Updated openstack/openstack Project: openstack-dev/devstack  d4da210ffb2a9f1274705ba04ec56337e7e1ea1c
Remove osapi_v3 configuration

Nova option osapi_v3 is used for Nova v2.1 API and the default value
was False(disabled). However Nova v2.1 API is CURRENT status and the
API should be enabled as the default as we discussed on
http://lists.openstack.org/pipermail/openstack-dev/2015-July/069624.html
We could not find it before because devstack makes it True, so this
patch removes it for avoiding confusion any more.

Change-Id: I4efd2036605a1a41ea297b44a5f31b2da7412593
Related-Bug: #1462901
Depends-on: I43f0352f9fa89401f79389a6dc1035d901f52ed2