Updated openstack/openstack Project: openstack/cinder  8e8f7f8016f4449888e2ce256974d3fbcbc51ec2
Backups: allow name to be specified during restore

This commit adds the ability to specify a `name` attribute during backup
restore. This will create a volume with the specified name during
restore, instead of the default pattern of the form
`restore-backup-<id>`. In case volume id is specified the `name`
attribute is ignored.

DocImpact
APIImpact
Implements: blueprint backup-restore-display-name
Change-Id: I23730834058d88e30be62624ada3b24cdaeaa6f3
Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@gmail.com>