Updated openstack/openstack Project: openstack/nova  9c5f39d45fefbb16cc1c42b181a9dc24240fccaa
Revert "Add config drive support for Virtuozzo containers"

This reverts commit def785a23d16e3965743bbb2a35ec5a3249a2fd0.

As discussed at the nova liberty mid-cycle, we still have concerns about
how this was implemented. We've also talked through an alternative
implementation (generate the config drive to a directory and then just map
that directory into the container).

I will work on the changes needed on the config drive site to implement
that now, and will help the PCS guys integrate that with their code.

Change-Id: I7106e9bf367e0d392566cf5271ae52fc1595284e