Updated openstack/openstack Project: openstack/horizon  7ca9b73f3632693926a0568fa3317c4051d1953a
Handling vnic type as optional

Currently there is no way to check vnci_type is supported or not from the
outside of neutron. If it's not supported, when trying to "edit port" a
Traceback was issued.
This patch checks if there is a vnci_type before accessing it in the
"edit port" view.

Change-Id: Ia3b75cb300832f01efbc933c3d5784af9b93f308
Closes-bug: #1477100