Updated openstack/openstack Project: openstack-dev/devstack  b488c81e031df361179a8ee73a70c4a6e859bbc4
Separate start/stop control of Neutron L2 agent.

This patch separates out control of the Neutron L2 agent from
starting/stopping the rest of Neutron.  This is needed for the same
reason that control of nova-compute was separated out for Nova.  When
doing rolling upgrade testing with Grenade, we need to be able to stop
and upgrade everything except the L2 agent, as that is what would be
running on a compute node.

After this is in place, we can update grenade to support a partial
upgrade scenario with Neutron and run it in jenkins to ensure we don't
break live upgrade support of Neutron.

Change-Id: I7eb87fba778aff3e4514813c6232dafa99ee2912
Signed-off-by: Russell Bryant <rbryant@redhat.com>