Updated openstack/openstack Project: openstack/python-keystoneclient  6363f989870cea59c6fcc60ae95aea383cfde0a9
Add get_token_data to token CRUD

We already have the validate method that returns an AccessInfo object.
For auth_token middleware it would be simpler if the client returned
simply the token data so it presented the same way as other sources of
token data.

It would also help with the keystoneauth transition in auth_token as we
could bypass the keystoneclient.AccessInfo objects.

Closes-Bug: #1475041
Change-Id: Ifbe7a7004937d910739c325cc04ae7264a4498e0