Updated openstack/openstack Project: openstack-infra/project-config  0147a8c53555d5d5e23c4566681a0c407c8f5ae5
Install dependencies needed for gems in macro

Some of the puppet jobs installed dependencies needed to compile
commonly used gems.  The infra puppet jobs lacked that.  Move that
logic into the 'uninstall-puppet' macro and rename it to 'puppet-
prepare-node' to reflect that it's a general macro for preparing
a node to run puppet jobs.

Apply the macro everywhere 'uninstall-puppet' was used before.
This adds the packaging logic to the infra jobs and some OpenStack
puppet jobs.

Change-Id: Ib504f3fea4713f14bf3dc2e9659fad0e999f8146