Updated openstack/openstack Project: openstack-infra/system-config  d98a7ab37ca0d0a53f77bf2c9c96e2d5de8ad91e
Switch jenkins.openstack.org to a snakeoil cert

The SSL cert for jenkins.openstack.org was obtained in 2013 when
this interface was more heavily used by our developer community.
Since then we've rolled out 7 additional Jenkins masters and so this
is no longer a useful primary reference. The only real need for
authenticated connections to its WebUI at this point is systems
administrators performing maintenance and minimal local
configuration, for which ToFU of a self-signed cert is sufficient.

Change-Id: Ibf95983a2ac76c2e9e39bcfc99643e3cac401245