Updated openstack/openstack Project: openstack-infra/puppet-graphite  ea59f65f75ba73abd46a77bc8a16c5c544b17885
Include mod_wsgi for graphite

If mod_wsgi is not installed, the puppet-graphite module will fail to
start apache because of the WSGI commands in the vhost template. This
patch adds the httpd::mod::wsgi class to take care of it. There is
nothing else in the graphite.openstack.org site.pp node that installs
mod_wsgi so the package resource in the class should not conflict with
anything.

Change-Id: Iaba16663c99a9a3aa6b69ac9b5af6dba22cd3fd2