Updated openstack/openstack Project: openstack-infra/project-config  08980584fe6aa23949ab887e0646e8a7a06abd9f
Start with a green field for RPM packaging

RDO and SUSE folks have agreed to start with a fresh
repo for collaborating on upstream packaging. In order to
sort out the issues, we start with a low hanging fruit
and iron out the differences there and will extend this
up the stack as things are progressing.

Depends-on: I608651414fadc0206910d2720d8d14be51a5c8af

Change-Id: Ibd1c082b6ae8d26e7fbeff42cd41643ba1b7768e