Updated openstack/openstack Project: openstack-infra/project-config  cad24c4b65615b37dd3b2a09a493a107ba5d42d8
Prep $PROJECT and upper constraints for install

Before we can start enforcing constraints during tox runs we need to
ensure that the upper constraints from the requirements repo is in place
and up to date with the correct ref which may be latest $BRANCH or
$ZUUL_REF. Use zuul-cloner to ensure the state of both the project under
test and the constraints from the requirements repo.

To make use of this projects can update their tox targets to look for an
upper constraints file in the root dir of the project and have pip honor
those constraints when present.

Change-Id: I2829ba1bb0ce22b0f18080a1dc4ca7b1c77ce0e3