Updated openstack/openstack Project: openstack/cinder  a3ec3dbc1f45999d4dc183bdba5de2185b52b128
Remove logging statements from migrations

One of the things that I believe is causing problems with the move to
ostestr and concurrency in unit test run is the copious amounts of
logging that is coming from the db migrations.  In most cases much of
this logging doesn't really seem all that useful, a debug or info log
stating a migration failed, followed by a raise which should generate
a stack trace with the error anyway, seems we could easily do without
this.

Also the "performing migration xxxx/completing migration xxxx" doesn't
seem overly helpful either.  Keep in mind that the migration engine
does some amount of logging anyway.

In addition, catching the exceptions and logging them isn't overly
beneficial or needed either.  When calling migrate from cinder-manage
if one of the migrations raises an exception that will be propogated
up to cinder-manage and displayed.  In other words all of the info is
already available, so there's no need to put it here.

Finally, if you take a look in general most migrations in other
projects don't LOG every thing for the same reasons.

So this patch removes logging statement from the migrations themeselves.

Change-Id: I285df7061c79c79e217cccb2fe387437c20ff746