Updated openstack/openstack Project: openstack/tempest  681a537cc1f7067dac172a2fee5d937000bf669f
Enable tempest install to setup a config dir

This commit adds the necessary bits to tempest's setup.cfg file to
instruct pbr on how to make sure that the sample config files and
a /etc/tempest dir get setup when installing tempest. In the new
cli workflow it is expected that by default /etc/tempest will exist
as a source for local working dir configs.

Partially implements: bp tempest-cli-improvements

Change-Id: I2bf93f6789601caaebdfb85ac2b9f34b82f87980