Updated openstack/openstack Project: openstack/os-cloud-config  0a52cd88ba122592843106b0c0912bf9637fb613
Use public ip for pki initialization

The host variable is used as the internal ip during endpoint creation.
Since the internal ip may not always be reachable from the machine
running os-cloud-config, we need to use the public variable (if provided)
instead when we ssh to the keystone host to do pki initialization.

Change-Id: I3850f0a1fca5f4560500aeee8ad9340338c89e77