Updated openstack/openstack Project: openstack/manila  3331f18a320c0c790bf0423af4dc0a217725dab0
Fix 'extend_share' in generic driver

When we create a share and mount it on client VM, we cannot do
'manila extend' because of error:
Stderr: u'umount: /shares/share-xxxxxx: device is busy.

We need to 'disable_access_for_maintenance' before doing an extension
and 'restore_access_after_maintenance' after extension was done.

Closes-Bug: #1473033

Change-Id: I4b5e665f8d028d4784d02411bbe6daace30aa49e