Updated openstack/openstack Project: openstack/neutron-vpnaas  33c3fd0302b1263291ece482b3b37a030a01d5c2
Support VPNaaS with L3 HA

VPNaaS with a HA router is not supported now.
This patch will enable support for VPNaaS with an HA router.

When VPN service is created for HA router, it should run
only on master node.
On transition from master to backup, vpn service should be
shutdown (same like disabling radvd) on that agent.

On transition from backup to master, vpn service should be
enabled and running on that agent.

Closes-bug: #1478012
Change-Id: I22f55b72cdc6cf608f50db902e4e3636fd59a16c