Updated openstack/openstack Project: openstack/ceilometer  0e586b940d834036a5e50988c30dc5c0b7bfcbb9
api: Redirect request to aodh if available

To ensure smooth migration to aodh we have to provide some backward
compatibilty. This will be done by redirecting people to the aodh API.

This will be done in two places:

- the ceilometerclient to allows to:
  * not query ceilometer-api if aodh present
  * allow to remove ceilometer-api if gnocchi+aodh is used
  see: https://review.openstack.org/#/c/202938/

- the ceilometer-api to:
  * allow old ceilometerclient to transparantly works
  * not break non-python client library

This change is about the ceilometer-api side.
It returns a 307 with the url of the aodh endpoint
if available to allow smooth transition to the aodh API.

APIImpact
DocImpact

Change-Id: I5216e3a70466447a796da6621f8495968e54a3c2