Updated openstack/openstack Project: openstack/requirements  630400718d8910f9da1bde1a20097076db022bcd
Increase dogpile.cache versio to support NullBackend

oslo.cache and other projects that use dogpile.cache
(like keystone) use custom version of backend that does nothing
(it is called NoOp backend). Instead of this it is better to use
dogpile realisation of such backend: NullBackend. It was
introduced in v0.5.4 so we need to increase it in global
requirements.

Change-Id: Idb6f78bf383c0f1d85dad33d9d34165e8b9c53ef