Updated openstack/openstack Project: openstack/oslo.config  ac505d83e4eeb7f4d0e813e4b33cf97d2d1613a2
Add enforce_type option when setting an override

The 'override' value of an option is currently stored and retrieved
with the exact value that is set through set_override method - not
taking into account the type of that option. Sometimes we want the type
of the override value to be the same as the option type and therefore
we will add an optional parameter to the set_override method that will
enforce the option value type.

Closes-Bug: #1461299
Change-Id: I008b76d3292f76d0699f0063930a3b190539740f