Updated openstack/openstack Project: openstack/heat  40f900005e6a2d67a4412a8a9932362cd47c8879
Fix RS CLB resource to handle HTTPS_REDIRECT correctly

When HTTPS_REDIRECT property for Rackspace cloud loadbalancer
is enabled, setting protocol to HTTP results in an error, even
if SSL_TERMINATION is specified correctly. This patch fixes
this behavior to allow the resource to complete when the above
options are correctly specified.

Change-Id: Ie8652835bcdf24f81205657d78811cc05337921e
Closes-Bug: #1476707