Updated openstack/openstack Project: openstack/oslo.config  257de47935487bf03555c37f6a42c74cb6476b0d
preserve preformatted help text

The config options are an interface to a subset of users. These users
need to read and understand the help text of the options. To make it
easier to read, it should be possible to preformat the text. This
patch set keeps the preformat with line breaks.

Change-Id: I8aa871001f7da3cefcd9aa93992d50d0d9140157