Updated openstack/openstack Project: openstack/horizon  746222bea76aa8c1cccd2ad19a245a8fa57c705e
Trying to get pool/health_monitor only if existing

If the filters to get health_monitor list or pool list are empty
list, all resources in that tenant will be returned. This leads
to issue described in the bug description. Fix this by trying
to retrieve pool/health_monitor list only if the filter is not
empty.

Change-Id: Ic45eea981cf752fffd84a8a1c930e2cdd6299bd8
Closes-bug: #1427459