Updated openstack/openstack Project: openstack/oslo.vmware  ddf834189b700acecea1698ff8102b759805f5c3
Make usage doc easier to understand for library consumers

This change makes a few modifications to make the usage doc easier
to understand and consume for users of the oslo.vmware library.
Among the changes are explanations for parameters, extra helpful
comments, removing references to default arguments to simplify the
example.

Change-Id: Iccbe40f6bda0c65c3de87d708634ae6bbdcaa959