Updated openstack/openstack Project: openstack/taskflow  b8ce6c0eaaae13711e0138bc52c391bcb4135025
Create and use a serial retry executor

To make it easily possible to change the retry
atom execution from being in the engine thread this
creates a retry executor (which is similar to the task
executor) and provide that a serial executor (which it will
use to execute with). This makes the retry and task actions
closer to being the same and makes the surrounding code that
much similar (which makes understanding it easier).

Change-Id: I993e938280df3bd97f8076293183ef21989e2dba