Updated openstack/openstack Project: openstack/oslo.log  8dee707e8ac1d620d07541fd5902c80ff46853a2
Define TRACE logging level

TRACE is a logging keyword that is understood by most logging
tools. OpenStack has repurposed this in the past to not be TRACE
logging but instead be used whenever a Stacktrace was dumped. This
was already fixed in:
I8c618e7b8523bcc0e19a9b77c4d2d5984369a04a

Now in this review, we are adding a new TRACE level set to an
integer halfway between DEBUG and NOTSET which is currently 5.

One can also use LOG.trace for deep tracing of their code as well.
To make it easy for end users to import the TRACE level, added
public constants for all the log levels from logging module. Several
projects import logging just for the constants, this will help
them as well.

Please see the logging guidelines for more info:
https://github.com/openstack/openstack-specs/blob/master/specs/log-guidelines.rst

Change-Id: Ica0e78da5a57c246b3b7499a6ce8c6f410cd6a80