Updated openstack/openstack Project: openstack/heat  502c16846e43d26151ea9b01a57a72755c90eca9
Stop using eventlet threads in test_engine_service

Currently we rely on actually running the tasks on in the
stack ThreadGroup, which will be non-deterministic where
wait() is not used (leading to some incorrect assertions)
and unneccesarily complex even when using wait().

So instead use a dummy implementation which allows for tasks
to be either run immediately or just recorded so we can
assert they would be started.

Change-Id: I2a748cac50e380289c33eadfa1c53309f86ad145
Partial-Bug: #1477484