Updated openstack/openstack Project: openstack/tempest  094ec36c28000f4550bdeccb53a0f123af498350
Use image minDisk as volume size when necessary

When creating a volume from an image, it is possible
that the configured volume_size value is less than
the minDisk of the image, which causes the test to
fail. In this case, the size of the created volume
should be the minDisk size instead.

Change-Id: I06f47fdcef7d36b9bb7793d29b5fe86bf6080b98
Closes-Bug: #1475746