Updated openstack/openstack Project: openstack/nova  25b378bf5c64269a57cf8a276ac3763b733032e2
Hyper-V: restart serial console workers after instance power change

At the moment, the only way the Hyper-V handles unexpected power
state changes is via the manager periodic task. The issue is that
unexpected instance reboots are not handled at all.

After an instance is stopped, suspended or paused, the console log
workers are stopped as well as the serial console pipe is closed.
Those workers are never restarted unless a power change is triggered
by the manager.

This patch proposes an instance state event listener, emiting events
that can be handled by the manager, similar to the Libvirt driver
workflow.

Change-Id: Ia0a875707a7a37224efb333feab1503e83449cc4
Closes-Bug: #1428505