Updated openstack/openstack Project: openstack/nova  a4115f21506868dc4d6d0320391d95773a4f9314
Add MKS protocol for remote consoles

MKS is the native protocol for VMware consoles and this patch is adding
API support for it. For now there is only one console type for this
protocol and this is 'webmks'. A new microversion is introduced which
adds protocol 'mks' and type 'webmks' for requesting a remote console.

Example request:

POST /servers/<uuid>/remote-consoles
{
  "remote_console": {
    "protocol": "mks",
    "type": "webmks"
  }
}

Example response:

{
  "remote_console": {
    "protocol": "mks",
    "type": "webmks",
    "url": "http://example.com:6090/mks.html?token=XYZ"
  }
}

APIImpact
Implements: blueprint vmware-webmks-console

Change-Id: I7217b999fb8d64a5646d8f20d7426c26553871d7