Updated openstack/openstack Project: openstack/heat  2ef0661e3a568899a8f97e1425131f3951621e76
Reduce size of stack_identity in heat notifications

Reduce "stack_identity" field in heat notifications because
it is too long for other consumers (at least for ceilometer).
So UUID of the stack will be used as stack identity instead
of arn in heat notifications.
NOTE: stack statistics for existing stacks will not be 
considered in ceilometer because we changed stack identity
and ceilometer recognize stacks with UUIDs as 
"new" resources. I consulted with gordonchung about this 
and it seems that it is not critical for ceilometer.

Closes-bug: #1468916
Change-Id: I74d192b8076c344dceab5754681c129a9e0fcd22